Run action on node rather than with Docker.
Change the action to be a pure JS action and run on node12. Running on
node means we can benefit from all the tools installed on GitHub runner
rather than what's provided with our own Docker image.

For example, running on `ubuntu-latest` means we have access to a very
recent (although not necessarily the latest) release of Gradle, plus
multiple jdk versions and the android sdk.