Fix git cli authorisation config.
Here we fix the git cli configuration to actually use the input repo-token
parameter: the token was previously used only in http requests to the GH
api, but it was being ignored by the git commands used to commit changes
and push to remote origin.

Using the same technique as the `checkout` action, we now save the token
credentials as a git local config (extraheader), so that an Authorization
header is correctly sent along each request to GitHub server. The config
is then removed in the post action.