Make sure wrapper jar is updated by running Gradle a second time.
By design, the `wrapper` task only updates the `gradle-wrapper.properties`
file with distribution url and checksum of the new target gradle release.
However, in order to update the wrapper jar itself (and the shell scripts)
it's necessary to run the `wrapper` task a second time.

See https://docs.gradle.org/current/userguide/gradle_wrapper.html#sec:upgrading_wrapper

As part of this change we're also starting to run the `wrapper` task
using the `gradlew` script available in the project rather than the one
installed in the environment. This should open the door to better support
for updating to RC releases (see #188).

Closes #233