Customize pull request title (#695)
* feat: Add `prTitleTemplate` input

This change adds a new input to the action. It can be used to customize
the title of the pull request created by this action. The same string is
also used in the pull request body.

This input is not required and has the default value Update Gradle
Wrapper from %sourceVersion% to %targetVersion%, where %sourceVersion%
will be replaced by the current/old version of the Gradle Wrapper and
%targetVersion% will be replaced by the new version of the Gradle
Wrapper.

Caveat: Since it's possible the source version of the Gradle Wrapper can
not be determined it may happen that the `%sourceVersion%` may be
replaced with the string 'undefined'. This could be handled by a second
new input to provide a fallback template for the case where the
`sourceVersion` is undefined.

The commit message remains unaffected.

* Update README.md

Co-authored-by: Cristian Greco <cristian@regolo.cc>

* chore: Document custom PR title fallback for undefined source version

* Update README.md

---------

Co-authored-by: Cristian Greco <cristian@regolo.cc>