Customize commit messages (#679)
* ignore .idea directory

* Add `commit-message-template` input

This change adds a new input to the action. It can be used to customize
the message of the commit created by this action.

This input is not required and has the default value `Update Gradle
Wrapper from %sourceVersion% to %targetVersion%`, where `%sourceVersion%`
will be replaced by the current/old version of the Gradle Wrapper and
`%targetVersion%` will be replaced by the new version of the Gradle
Wrapper.

The PR title and body remain unaffected by this change.

Implements gradle-update/update-gradle-wrapper-action#246

* Align behaviour for `undefined` sourceVersion with #695

For cases where the source version of the Gradle wrapper cannot
be determined the string `undefined` is used as a fallback.

This allows to use the same custom commit message template in
all cases.

* Update action.yml

---------

Co-authored-by: Cristian Greco <cristian@regolo.cc>